engine-1-leadgen/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
starlette>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
itsdangerous>=2.1.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
